requirements.txt
appium-python-client
pytest
allure-python-commons
python-dotenv
selenium
